Return 64 bit integers from all SIMD spline evaluations

+++ spline/interp2Dcomp.c
@@ -111,7 +111,7 @@ int interp2Dcomp_init(interp2D_data* str, real* f, int n_r, int n_z,
  * @param r r-coordinate
  * @param z z-coordinate
  */
-int interp2Dcomp_eval_B(real* B, interp2D_data* str, real r, real z) {
+integer interp2Dcomp_eval_B(real* B, interp2D_data* str, real r, real z) {
     int i_r = (r-str->r_min)/str->r_grid; /**< index for r variable */
     real dr = (r-(str->r_min+i_r*str->r_grid))/str->r_grid; /**< Normalized r coordinate in
 							       current cell */
@@ -154,7 +154,7 @@ int interp2Dcomp_eval_B(real* B, interp2D_data* str, real r, real z) {
     return err;
 }
 
-int interp2Dcomp_eval_B_SIMD(int i, real B[NSIMD], interp2D_data* str, real r, real z) {
+integer interp2Dcomp_eval_B_SIMD(int i, real B[NSIMD], interp2D_data* str, real r, real z) {
     int i_r = (r-str->r_min)/str->r_grid; /**< index for r variable */
     real dr = (r-(str->r_min+i_r*str->r_grid))/str->r_grid; /**< Normalized r coordinate in
 							       current cell */
@@ -212,7 +212,7 @@ int interp2Dcomp_eval_B_SIMD(int i, real B[NSIMD], interp2D_data* str, real r, r
  * @param r r-coordinate
  * @param z z-coordinate
  */
-long int interp2Dcomp_eval_dB(real* B_dB, interp2D_data* str, real r, real z) {
+integer interp2Dcomp_eval_dB(real* B_dB, interp2D_data* str, real r, real z) {
     int i_r = (r-str->r_min)/str->r_grid;                   /**< index for r variable */
     real dr = (r-(str->r_min+i_r*str->r_grid))/str->r_grid; /**< Normalized r coordinate in
 							       current cell */
@@ -322,7 +322,7 @@ long int interp2Dcomp_eval_dB(real* B_dB, interp2D_data* str, real r, real z) {
     return err;
 }
 
-int interp2Dcomp_eval_dB_SIMD(int i, real B_dB[6][NSIMD], interp2D_data* str, real r, real z) {
+integer interp2Dcomp_eval_dB_SIMD(int i, real B_dB[6][NSIMD], interp2D_data* str, real r, real z) {
     int i_r = (r-str->r_min)/str->r_grid;                   /**< index for r variable */
     real dr = (r-(str->r_min+i_r*str->r_grid))/str->r_grid; /**< Normalized r coordinate in
 							       current cell */
